+++ gui/freeadmin/static/lib/js/dgrid/CHANGES.md
@@ -5,15 +5,45 @@ This document outlines changes since 0.3.0.  For older changelogs, see the
 
 ## Significant changes
 
+### General/Core
+
 * The `up` and `down` methods of `List` will now call `grid.row` internally to
     resolve whatever argument is passed; the `left` and `right` methods of
     `Grid` will call `grid.cell`.  (Formerly these methods only accepted a
     row or cell object directly.)
+* The `Grid` module now emits a `dgrid-sort` event when a sortable header cell
+    is clicked; this event includes a `sort` property, and may be canceled to
+    stop the sort, or to substitute alternative behavior.  In the latter case,
+    if updating the sort arrow in the UI is still desired, call the
+    `updateSortArrow` method and pass the `sort` value from the event.
+* The `OnDemandList` module now supports a `pagingMethod` property, which allows
+    specifying whether to throttle or debounce scroll events.  The default
+    behavior has been changed from `"throttleDelayed"` to `"debounce"`, which
+    generally is capable of far reducing the number of store queries issued,
+    moreso if `pagingDelay` is increased (though its default remains the same).
+* The `OnDemandList` module now supports a `keepScrollPosition` property, which
+    will attempt to preserve scroll position between refresh calls.  This can be
+    set on the instance itself to affect all refreshes, or can be passed to the
+    `refresh` method directly for a specific call.
+* The `OnDemandList` module now returns a promise from the `refresh` method,
+    which resolves when the grid finishes rendering results after the refresh.
+    It also emits a `dgrid-refresh-complete` event, which includes both a
+    reference to the QueryResults object (`results`) and the rendered rows
+    (`rows`).
+
+### Extensions
+
+* The `Pagination` extension now returns a promise from the `refresh` and
+    `gotoPage` methods, which resolves when the grid finishes rendering results.
+    Note that it does not (yet) emit an event like `OnDemandList`.
 
 ## Other changes and fixes
 
 ### General/Core
 
+* Resolved an issue where upon changing column structure, the placement of the
+    sort arrow would be lost even though the grid is still sorting by the same
+    field.
 * Resolved an issue where OnDemandList could end up firing requests where
     start exceeds total and count is negative. (#323)